+++ test/positioning.spec.js
@@ -40,18 +40,34 @@ test("should correctly calculate position related to parent", () => {
 	const testCases = [
 		["bottom", parent, { x: 399, y: 551, z: 20, width: "auto" }],
 		["top", parent, { x: 399, y: 300, z: 20, width: "auto" }],
-		["left", parent, { x: 300, y: 400, z: 20, width: "auto" }],
+		["left", parent, { x: 299, y: 400, z: 20, width: "auto" }],
 		["right", parent, { x: 551, y: 400, z: 20, width: "auto" }],
 		["bottom-right", parent, { x: 551, y: 551, z: 20, width: "auto" }],
-		["bottom-left", parent, { x: 300, y: 551, z: 20, width: "auto" }],
+		["bottom-left", parent, { x: 299, y: 551, z: 20, width: "auto" }],
 		["top-right", parent, { x: 551, y: 300, z: 20, width: "auto" }],
-		["top-left", parent, { x: 300, y: 300, z: 20, width: "auto" }],
+		["top-left", parent, { x: 299, y: 300, z: 20, width: "auto" }],
 
 		["bottom-fit", parent, { x: 399, y: 551, z: 20, width: "150px" }],
 		["top-fit", parent, { x: 399, y: 300, z: 20, width: "150px" }],
 
 		["center", container, { x: 650, y: 650, z: 20, width: "auto" }],
 		["center-fit", container, { x: 0, y: 650, z: 20, width: "1400px" }],
+
+		["left-start", parent, { x: 299, y: 400, z: 20, width: "auto" }],
+		["left-center", parent, { x: 299, y: 425, z: 20, width: "auto" }],
+		["left-end", parent, { x: 299, y: 450, z: 20, width: "auto" }],
+
+		["right-start", parent, { x: 551, y: 400, z: 20, width: "auto" }],
+		["right-center", parent, { x: 551, y: 425, z: 20, width: "auto" }],
+		["right-end", parent, { x: 551, y: 450, z: 20, width: "auto" }],
+
+		["top-start", parent, { x: 400, y: 300, z: 20, width: "auto" }],
+		["top-center", parent, { x: 425, y: 300, z: 20, width: "auto" }],
+		["top-end", parent, { x: 450, y: 300, z: 20, width: "auto" }],
+
+		["bottom-start", parent, { x: 400, y: 551, z: 20, width: "auto" }],
+		["bottom-center", parent, { x: 425, y: 551, z: 20, width: "auto" }],
+		["bottom-end", parent, { x: 450, y: 551, z: 20, width: "auto" }],
 	];
 
 	testCases.forEach(([at, parent, expected], i) => {
@@ -77,4 +93,120 @@ test("should correctly calculate position at point if parent is bigger than cont
 	const expected = { x: 200, y: 240, z: 20, width: "auto" };
 	const result = calculatePosition(portal, parent, "point", left, top);
 	expect(result).toEqual(expected);
+});
+
+test("should change 'at' position when there is not enough space between parent and container", () => {
+	document.body.style = "margin: 0; padding: 0;";
+	const style =
+		"padding-top: 200px; padding-left: 200px; box-sizing: border-box;";
+	document.body.innerHTML = `
+	    <div id="container" style="position: relative; width: 710px; height: 500px; ${style}">
+	    	<div id="parentCont" style="position: relative; width: 300px; height: 300px;">
+	    		<div id="parent" style="width: 200px; height: 50px; "></div>
+				</div>
+				<div id="popup" style="position: absolute; width: 100px; height: 100px;"></div>
+			</div>
+	`;
+	const popup = document.getElementById("popup");
+	const parent = document.getElementById("parent");
+	const parentCont = document.getElementById("parentCont");
+
+	const testCases = [
+		["top", { x: 200, y: -200 }, { x: 399, y: 50, z: 20, width: "auto" }],
+		["top-right", { x: 200, y: -200 }, { x: 601, y: 0, z: 20, width: "auto" }],
+		["top-left", { x: 200, y: -200 }, { x: 299, y: 0, z: 20, width: "auto" }],
+		["top-right", { x: 250, y: -200 }, { x: 610, y: 50, z: 20, width: "auto" }],
+		["top-left", { x: -200, y: -200 }, { x: 0, y: 50, z: 20, width: "auto" }],
+		["top-end", { x: -200, y: -200 }, { x: 100, y: 50, z: 20, width: "auto" }],
+		[
+			"top-center",
+			{ x: -200, y: -200 },
+			{ x: 50, y: 50, z: 20, width: "auto" },
+		],
+		["left-end", { x: -200, y: -200 }, { x: 200, y: 0, z: 20, width: "auto" }],
+		["left-end", { x: 250, y: -200 }, { x: 349, y: 0, z: 20, width: "auto" }],
+		["right-end", { x: 250, y: -200 }, { x: 350, y: 0, z: 20, width: "auto" }],
+
+		["bottom", { x: 200, y: 200 }, { x: 399, y: 300, z: 20, width: "auto" }],
+		[
+			"bottom-start",
+			{ x: 200, y: 200 },
+			{ x: 400, y: 300, z: 20, width: "auto" },
+		],
+		[
+			"right-start",
+			{ x: 250, y: 250 },
+			{ x: 350, y: 400, z: 20, width: "auto" },
+		],
+		["bottom", { x: -200, y: 200 }, { x: 0, y: 300, z: 20, width: "auto" }],
+		[
+			"bottom-left",
+			{ x: -200, y: 200 },
+			{ x: 0, y: 300, z: 20, width: "auto" },
+		],
+		[
+			"left-start",
+			{ x: -200, y: 250 },
+			{ x: 200, y: 400, z: 20, width: "auto" },
+		],
+		["left-bottom", { x: -99, y: 250 }, { x: 0, y: 400, z: 20, width: "auto" }],
+		[
+			"left-bottom",
+			{ x: -150, y: 250 },
+			{ x: 0, y: 350, z: 20, width: "auto" },
+		],
+		[
+			"right-bottom",
+			{ x: 150, y: 250 },
+			{ x: 551, y: 400, z: 20, width: "auto" },
+		],
+		[
+			"right-bottom",
+			{ x: 250, y: 250 },
+			{ x: 610, y: 350, z: 20, width: "auto" },
+		],
+	];
+
+	testCases.forEach(([at, offset, expected], i) => {
+		parentCont.style.left = offset.x + "px";
+		parentCont.style.top = offset.y + "px";
+		const result = calculatePosition(popup, parent, at);
+		expect(result, `#${i} ${at}`).toEqual(expected);
+	});
+});
+
+test("should correctly apply 'at' position when container styles contain borders", () => {
+	document.body.style = "margin: 0; padding: 0;";
+	const style = "box-sizing: border-box; border: 10px solid #eee;";
+	document.body.innerHTML = `
+	    <div id="container" style="position: relative; width: 500px; height: 500px; ${style}">
+	    	<div id="parentCont" style="position: relative; width: 300px; height: 300px;">
+	    		<div id="parent" style="width: 200px; height: 50px; "></div>
+				</div>
+				<div id="popup" style="position: absolute; width: 100px; height: 100px;"></div>
+			</div>
+	`;
+	const popup = document.getElementById("popup");
+	const parent = document.getElementById("parent");
+	const parentCont = document.getElementById("parentCont");
+
+	const testCases = [
+		[
+			"bottom-right",
+			{ x: 200, y: 200 },
+			{ x: 380, y: 251, z: 20, width: "auto" },
+		],
+		[
+			"bottom-right",
+			{ x: 150, y: 340 },
+			{ x: 351, y: 380, z: 20, width: "auto" },
+		],
+	];
+
+	testCases.forEach(([at, offset, expected], i) => {
+		parentCont.style.left = offset.x + "px";
+		parentCont.style.top = offset.y + "px";
+		const result = calculatePosition(popup, parent, at);
+		expect(result, `#${i} ${at}`).toEqual(expected);
+	});
 });